some clean code, and new router mounted
