+++ src/routers/permission/add_permission.rs
@@ -1,22 +1,12 @@
 use crate::{
-    authz::{Condition, ConditionValueType, ModelAttrib, ModelAttribResult},
     error::RouterError,
     models::{NewPermission, NewPermissionCondition, Permission},
     DbPool,
 };
 use actix_web::web;
 use diesel::prelude::*;
-use serde::{Deserialize, Serialize};
 
-use super::SimpleCondition;
-
-#[derive(Serialize, Deserialize)]
-pub struct NewPermissionData {
-    subject: String,
-    object: String,
-    action: String,
-    conditions: Vec<SimpleCondition>,
-}
+use super::NewPermissionData;
 
 pub async fn add_permission<'a>(
     new_permission: web::Json<NewPermissionData>,
@@ -45,17 +35,7 @@ pub async fn add_permission<'a>(
         let mut insertable_conditions: Vec<NewPermissionCondition> = Vec::new();
 
         for condition in new_permission_data.conditions {
-            let model_attr = ModelAttrib::try_from(condition.name.as_str())?;
-            let attr_result = ModelAttribResult::from(model_attr);
-            let value_type = attr_result.get_value_type();
-
-            let request_value_type = ConditionValueType::try_from(condition.value.as_str())?;
-
-            if value_type != request_value_type {
-                return Err(RouterError::BadRequest(
-                    "Condition value type is not correct!".to_string(),
-                ));
-            }
+            let _ = condition.validate()?;
 
             insertable_conditions.push(NewPermissionCondition {
                 permission_id: new_permission.id,
@@ -68,10 +48,8 @@ pub async fn add_permission<'a>(
             .insert_into(app_permission_conditions)
             .execute(&mut conn)?;
 
-        Ok(())
+        Ok("Added")
     })
     .await
-    .unwrap()?;
-
-    Ok("Added")
+    .unwrap()
 }
